Create a dibiously functional adaptive stepsize alogrithm
I hadn't commited this code for some time, but it contains over a week of work in developing and testing the multistep algorithm. The results are mixed; while the model can adaptively step it still is not stable enough nor fast enough to be confident in its results. After this commit, I will attempt to incorporate a temporary version of the optimization work done by Carol on the other branch.